+++ crates/nox-tests/tests/spells.rs
@@ -29,26 +29,34 @@ use connected_client::ConnectedClient;
 use created_swarm::system_services_config::{DeciderConfig, SystemServicesConfig};
 use created_swarm::{make_swarms, make_swarms_with_cfg};
 use fluence_spell_dtos::trigger_config::{ClockConfig, TriggerConfig};
+use hex::FromHex;
 use log_utils::enable_logs;
 use service_modules::load_module;
 use spell_event_bus::api::{TriggerInfo, TriggerInfoAqua, MAX_PERIOD_SEC};
 use test_utils::{create_service, create_service_worker};
+use workers::CUID;
 
 type SpellId = String;
 type WorkerPeerId = String;
 
 async fn create_worker(client: &mut ConnectedClient, deal_id: Option<String>) -> WorkerPeerId {
+    let init_id_1 =
+        <CUID>::from_hex("54ae1b506c260367a054f80800a545f23e32c6bc4a8908c9a794cb8dad23e5ea")
+            .unwrap();
+    let unit_ids = vec![init_id_1];
     let data = hashmap! {
         "deal_id" => deal_id.map(JValue::String).unwrap_or(JValue::String("default_deal".to_string())),
         "relay" => json!(client.node.to_string()),
         "client" => json!(client.peer_id.to_string()),
+        "cu_ids" => json!(unit_ids)
     };
+
     let response = client
         .execute_particle(
             r#"
             (seq
                 (xor
-                    (call relay ("worker" "create") [deal_id] worker_peer_id)
+                    (call relay ("worker" "create") [deal_id cu_ids] worker_peer_id)
                     (seq
                         (call relay ("worker" "get_worker_id") [deal_id] get_worker_peer_id)
                         (ap get_worker_peer_id.$.[0] worker_peer_id)
@@ -172,7 +180,7 @@ async fn spell_simple_test() {
             (seq
                 (seq
                     (call %init_peer_id% (spell_id "get_script") [] script)
-                    (call %init_peer_id% (spell_id "get_u32") ["counter"] counter)
+                    (call %init_peer_id% (spell_id "get_u32") ["hw_counter"] counter)
                 )
                 (call "{}" ("return" "") [script.$.value counter])
             )
@@ -394,7 +402,7 @@ async fn spell_run_oneshot() {
         (seq
             (seq
                 (call relay ("op" "noop") [])
-                (call worker (spell_id "get_u32") ["counter"] counter)
+                (call worker (spell_id "get_u32") ["hw_counter"] counter)
             )
             (call client ("return" "") [counter])
         )"#,
@@ -436,7 +444,7 @@ async fn spell_install_ok_empty_config() {
         (seq
             (seq
                 (call relay ("op" "noop") [])
-                (call worker (spell_id "get_u32") ["counter"] counter)
+                (call worker (spell_id "get_u32") ["hw_counter"] counter)
             )
             (call %init_peer_id% ("return" "") [counter])
         )"#,
@@ -471,7 +479,7 @@ async fn spell_install_ok_empty_config() {
         (seq
             (seq
                 (call relay ("op" "noop") [])
-                (call worker (spell_id "get_u32") ["counter"] counter)
+                (call worker (spell_id "get_u32") ["hw_counter"] counter)
             )
             (call %init_peer_id% ("return" "") [counter])
         )"#,
@@ -896,7 +904,7 @@ async fn spell_call_by_alias() {
             (seq
                 (seq
                     (call %init_peer_id% ("srv" "add_alias") ["alias" spell_id])
-                    (call %init_peer_id% ("alias" "get_u32") ["counter"] counter)
+                    (call %init_peer_id% ("alias" "get_u32") ["hw_counter"] counter)
                 )
 
                 (call "{}" ("return" "") [counter.$.value])
@@ -932,7 +940,7 @@ async fn spell_trigger_connection_pool() {
         (seq
             (seq
                 (call %init_peer_id% ("getDataSrv" "spell_id") [] spell_id)
-                (call %init_peer_id% (spell_id "get_u32") ["counter"] counter)
+                (call %init_peer_id% (spell_id "get_u32") ["hw_counter"] counter)
             )
             (call "{}" ("return" "") [spell_id])
         )
@@ -1018,7 +1026,7 @@ async fn spell_timer_trigger_mailbox_test() {
         (seq
             (seq
                 (call %init_peer_id% ("getDataSrv" "spell_id") [] spell_id)
-                (call %init_peer_id% ("getDataSrv" "trigger") [] trigger)
+                (call %init_peer_id% ("getDataSrv" "hw_trigger") [] trigger)
             )
             (call "{}" ("return" "") [trigger])
         )
@@ -1059,7 +1067,7 @@ async fn spell_connection_pool_trigger_test() {
                 (seq
                     (call %init_peer_id% ("getDataSrv" "spell_id") [] spell_id)
                     (seq
-                        (call %init_peer_id% ("getDataSrv" "trigger") [] trigger)
+                        (call %init_peer_id% ("getDataSrv" "hw_trigger") [] trigger)
                         (call %init_peer_id% ("run-console" "print") ["getDataSrv, trigger:" trigger])
                     )
                 )
@@ -1119,56 +1127,51 @@ async fn spell_connection_pool_trigger_test() {
 
 #[tokio::test]
 async fn spell_set_u32() {
+    enable_logs();
     let swarms = make_swarms(1).await;
-    let mut client = ConnectedClient::connect_to(swarms[0].multiaddr.clone())
-        .await
-        .wrap_err("connect client")
-        .unwrap();
+    let mut client = ConnectedClient::connect_with_keypair(
+        swarms[0].multiaddr.clone(),
+        Some(swarms[0].management_keypair.clone()),
+    )
+    .await
+    .wrap_err("connect client")
+    .unwrap();
 
-    let script = format!(r#"(call "{}" ("return" "") ["called"])"#, client.peer_id);
-    let mut config = TriggerConfig::default();
-    config.connections.connect = true;
-
-    let (spell_id, worker_id) =
-        create_spell(&mut client, &script, config.clone(), json!({}), None).await;
-
-    let data = hashmap! {
-        "spell_id" => json!(spell_id),
-        "worker" => json!(worker_id),
-        "relay" => json!(client.node.to_string()),
-        "client" => json!(client.peer_id.to_string()),
-        "config" => json!(config),
-    };
-    let mut result = client
-        .execute_particle(
-            r#"(seq
+    let script = format!(
+        r#"( seq
+        (seq
+            (call %init_peer_id% ("getDataSrv" "spell_id") [] spell_id)
             (seq
+                (call %init_peer_id% (spell_id "get_u32") ["test"] absent)
                 (seq
-                    (call relay ("op" "noop") [])
-                    (call worker (spell_id "get_u32") ["test"] absent)
-                )
-                (seq
-                    (call worker (spell_id "set_u32") ["test" 1])
+                    (call %init_peer_id% (spell_id "set_u32") ["test" 1])
                     (seq
-                        (call worker (spell_id "get_u32") ["test"] one)
+                        (call %init_peer_id% (spell_id "get_u32") ["test"] one)
                         (seq
-                            (call worker (spell_id "set_u32") ["test" 2])
-                            (call worker (spell_id "get_u32") ["test"] two)
+                            (call %init_peer_id% (spell_id "set_u32") ["test" 2])
+                            (call %init_peer_id% (spell_id "get_u32") ["test"] two)
                         )
                     )
                 )
             )
-            (call %init_peer_id% ("return" "") [absent one two])
-           )"#,
-            data,
         )
-        .await
-        .unwrap();
+        (call "{}" ("return" "") [absent one two])
+    )"#,
+        client.peer_id
+    );
+    // oneshot spell
+    let mut config = TriggerConfig::default();
+    config.clock.start_sec = 1;
+    config.clock.period_sec = 0;
+
+    let (_spell_id, _worker_id) =
+        create_spell(&mut client, &script, config.clone(), json!({}), None).await;
+    let mut result = client.receive_args().await.wrap_err("receive").unwrap();
+
     assert_eq!(result.len(), 3);
     let (absent, one, two) = (result.remove(0), result.remove(0), result.remove(0));
 
     assert_eq!(absent["absent"], json!(true));
-
     assert_eq!(one["absent"], json!(false));
     assert_eq!(one["value"], json!(1));
 
@@ -1213,7 +1216,7 @@ async fn spell_update_config() {
         r#"(seq
             (seq
                 (call %init_peer_id% ("getDataSrv" "spell_id") [] spell_id)
-                (call %init_peer_id% ("getDataSrv" "trigger") [] trigger)
+                (call %init_peer_id% ("getDataSrv" "hw_trigger") [] trigger)
              )
             (call "{}" ("return" "") [trigger])
         )"#,
@@ -1303,7 +1306,7 @@ async fn spell_update_config_stopped_spell() {
         r#"(seq
             (seq
                 (call %init_peer_id% ("getDataSrv" "spell_id") [] spell_id)
-                (call %init_peer_id% ("getDataSrv" "trigger") [] trigger)
+                (call %init_peer_id% ("getDataSrv" "hw_trigger") [] trigger)
              )
             (call "{}" ("return" "") [trigger])
         )"#,
@@ -1504,10 +1507,14 @@ async fn spell_create_worker_twice() {
         .await
         .wrap_err("connect client")
         .unwrap();
-
+    let init_id_1 =
+        <CUID>::from_hex("54ae1b506c260367a054f80800a545f23e32c6bc4a8908c9a794cb8dad23e5ea")
+            .unwrap();
+    let unit_ids = vec![init_id_1];
     let data = hashmap! {
         "client" => json!(client.peer_id.to_string()),
         "relay" => json!(client.node.to_string()),
+        "cu_ids" => json!(unit_ids)
     };
     client
         .send_particle(
@@ -1515,11 +1522,11 @@ async fn spell_create_worker_twice() {
         (xor
             (seq
                 (seq
-                    (call relay ("worker" "create") ["deal_id"] worker_peer_id)
+                    (call relay ("worker" "create") ["deal_id" cu_ids] worker_peer_id)
                     (call relay ("worker" "get_worker_id") ["deal_id"] get_worker_peer_id)
                 )
                 (seq
-                    (call relay ("worker" "create") ["deal_id"] failed_create)
+                    (call relay ("worker" "create") ["deal_id" cu_ids] failed_create)
                     (call client ("return" "") ["test failed"])
                 )
             )
@@ -1593,16 +1600,20 @@ async fn spell_create_worker_same_deal_id_different_peer() {
         .await
         .wrap_err("connect client")
         .unwrap();
-
+    let init_id_1 =
+        <CUID>::from_hex("54ae1b506c260367a054f80800a545f23e32c6bc4a8908c9a794cb8dad23e5ea")
+            .unwrap();
+    let unit_ids = vec![init_id_1];
     let data = hashmap! {
         "client" => json!(client1.peer_id.to_string()),
         "relay" => json!(client1.node.to_string()),
+        "cu_ids" => json!(unit_ids)
     };
     let response = client1
         .execute_particle(
             r#"
         (seq
-            (call relay ("worker" "create") ["deal_id"] worker_peer_id)
+            (call relay ("worker" "create") ["deal_id" cu_ids] worker_peer_id)
             (call client ("return" "") [worker_peer_id])
         )"#,
             data.clone(),
@@ -1616,13 +1627,14 @@ async fn spell_create_worker_same_deal_id_different_peer() {
     let data = hashmap! {
         "client" => json!(client2.peer_id.to_string()),
         "relay" => json!(client2.node.to_string()),
+           "cu_ids" => json!(unit_ids)
     };
     let response = client2
         .execute_particle(
             r#"
         (xor
             (seq
-                (call relay ("worker" "create") ["deal_id"] worker_peer_id)
+                (call relay ("worker" "create") ["deal_id" cu_ids] worker_peer_id)
                 (call client ("return" "") ["test_failed"])
             )
             (call client ("return" "") [%last_error%.$.message])
@@ -1731,7 +1743,7 @@ async fn spell_update_trigger_by_alias() {
         (seq
             (seq
                 (call %init_peer_id% ("getDataSrv" "spell_id") [] spell_id)
-                (call %init_peer_id% (spell_id "get_u32") ["counter"] counter)
+                (call %init_peer_id% (spell_id "get_u32") ["hw_counter"] counter)
             )
             (call "{}" ("return" "") [spell_id])
         )
@@ -1942,7 +1954,7 @@ async fn spell_call_by_default_alias() {
         (seq
             (seq
                 (seq
-                    (call %init_peer_id% ("spell" "get_u32") ["counter"] counter1)
+                    (call %init_peer_id% ("spell" "get_u32") ["hw_counter"] counter1)
                     (seq
                         (call %init_peer_id% ("srv" "resolve_alias") ["spell"] spell_id1)
                         (xor
@@ -1952,7 +1964,7 @@ async fn spell_call_by_default_alias() {
                     )
                 )
                 (seq
-                    (call %init_peer_id% ("self" "get_u32") ["counter"] counter2)
+                    (call %init_peer_id% ("self" "get_u32") ["hw_counter"] counter2)
                     (seq
                         (call %init_peer_id% ("srv" "resolve_alias") ["self"] spell_id2)
                         (xor
@@ -1999,9 +2011,14 @@ async fn get_worker_peer_id_opt() {
         .wrap_err("connect client")
         .unwrap();
 
+    let init_id_1 =
+        <CUID>::from_hex("54ae1b506c260367a054f80800a545f23e32c6bc4a8908c9a794cb8dad23e5ea")
+            .unwrap();
+    let unit_ids = vec![init_id_1];
     let data = hashmap! {
         "relay" => json!(client.node.to_string()),
         "client" => json!(client.peer_id.to_string()),
+        "cu_ids"=>json!(unit_ids)
     };
     let response = client
         .execute_particle(
@@ -2009,7 +2026,7 @@ async fn get_worker_peer_id_opt() {
             (seq
                 (seq
                     (call relay ("worker" "get_worker_id") ["deal_id"] worker_peer_id_before)
-                    (call relay ("worker" "create") ["deal_id"] worker_peer_id)
+                    (call relay ("worker" "create") ["deal_id" cu_ids] worker_peer_id)
                 )
                 (seq
                     (call relay ("worker" "get_worker_id") ["deal_id"] worker_peer_id_after)
@@ -2252,7 +2269,7 @@ async fn test_activate_deactivate() {
 
     client
         .send_particle(
-    r#"(seq
+            r#"(seq
                 (seq
                     (seq
                         (call relay ("worker" "is_active") [deal_id] is_active_before)