+++ crates/subnet-resolver/src/resolve.rs
@@ -21,15 +21,13 @@ pub fn parse_chain_data(data: &str) -> Result<Vec<Token>, ChainDataError> {
     let signature: ParamType = Array(Box::new(Tuple(vec![
         // bytes32 id
         FixedBytes(32),
-        // bytes32 peerId
-        FixedBytes(32),
         // bytes32 workerId
         FixedBytes(32),
-        // address owner
+        // bytes32 peerId
+        FixedBytes(32),
+        // address provider
         Address,
-        // uint256 collateral
-        Uint(256),
-        // uint256 created
+        // uint256 joinedEpoch
         Uint(256),
     ])));
     Ok(ethabi::decode(&[signature], &data)?)
@@ -64,9 +62,6 @@ fn decode_pats(data: String) -> Result<Vec<Worker>, ResolveSubnetError> {
         let pat_id = next_opt(&mut tuple, "pat_id", Token::into_fixed_bytes)?;
         let pat_id = hex::encode(pat_id);
 
-        let peer_id = next_opt(&mut tuple, "compute_peer_id", Token::into_fixed_bytes)?;
-        let peer_id = parse_peer_id(peer_id)
-            .map_err(|e| ResolveSubnetError::InvalidPeerId(e, "compute_peer_id"))?;
         let worker_id = next_opt(&mut tuple, "compute_worker_id", Token::into_fixed_bytes)?;
         // if all bytes are 0, then worker_id is considered empty
         let all_zeros = worker_id.iter().all(|b| *b == 0);
@@ -77,6 +72,10 @@ fn decode_pats(data: String) -> Result<Vec<Worker>, ResolveSubnetError> {
                 .map_err(|e| ResolveSubnetError::InvalidPeerId(e, "worker_id"))?;
             vec![worker_id.to_string()]
         };
+
+        let peer_id = next_opt(&mut tuple, "compute_peer_id", Token::into_fixed_bytes)?;
+        let peer_id = parse_peer_id(peer_id)
+            .map_err(|e| ResolveSubnetError::InvalidPeerId(e, "compute_peer_id"))?;
 
         let pat = Worker {
             pat_id: format!("0x{}", pat_id),