+++ crates/workers/src/lib.rs
@@ -8,6 +8,10 @@ mod persistence;
 mod scope;
 mod workers;
 
+pub use core_manager::manager::CoreManager;
+pub use core_manager::manager::DummyCoreManager;
+pub use core_manager::manager::PersistentCoreManager;
+pub use core_manager::CUID;
 pub use error::KeyStorageError;
 pub use error::WorkersError;
 pub use key_storage::KeyStorage;